Refactor custom routing
