+++ src/routing/application/apply.route.tsx
@@ -3,7 +3,7 @@ import { Route, Switch, useRouteMatch } from 'react-router-dom';
 import { ApplicationTypeEnum } from '../../models/enums/application-type';
 import { Error404, PageProps } from '../../pages';
 import ApplyPage from '../../pages/Application/ApplyPage';
-import RestrictedRoute from '../route.extensions';
+import RestrictedRoute from '../RestrictedRoute';
 
 interface Props extends PageProps {
   type: ApplicationTypeEnum;
@@ -13,9 +13,14 @@ const ApplyRoute: FC<Props> = ({ paths, type }) => {
   const { path } = useRouteMatch();
   return (
     <Switch>
-      <RestrictedRoute path={`${path}`}>
-        <ApplyPage paths={paths} type={type} />
-      </RestrictedRoute>
+      <Route
+        path={`${path}`}
+        render={() => (
+          <RestrictedRoute>
+            <ApplyPage paths={paths} type={type} />
+          </RestrictedRoute>
+        )}
+      />
       <Route path="*">
         <Error404 />
       </Route>
