+++ src/routing/identity/LoginRoute.tsx
@@ -4,7 +4,7 @@ import { useQueryParams } from '../../hooks';
 import { LOCAL_STORAGE_RETURN_URL_KEY, RETURN_URL } from '../../models/constants';
 import LoginPage from '../../pages/Authentication/LoginPage';
 import LoginSuccessPage from '../../pages/Authentication/LoginSuccessPage';
-import { NotAuthenticatedRoute } from '../route.extensions';
+import { NotAuthenticatedRoute } from '../NotAuthenticatedRoute';
 
 export const LoginRoute: FC = () => {
   const params = useQueryParams();
@@ -19,9 +19,16 @@ export const LoginRoute: FC = () => {
 
   return (
     <Switch>
-      <NotAuthenticatedRoute exact path={`${path}`}>
-        <LoginPage flow={flow} />
-      </NotAuthenticatedRoute>
+      <Route
+        exact
+        path={`${path}`}
+        render={() => (
+          <NotAuthenticatedRoute exact path={`${path}`}>
+            <LoginPage flow={flow} />
+          </NotAuthenticatedRoute>
+        )}
+      />
+
       <Route exact path={`${path}/success`}>
         <LoginSuccessPage />
       </Route>
